+++ src/components/pages/DaoSettings/DAOSettings.tsx
@@ -3,13 +3,13 @@ import { useFractal } from '../../../providers/App/AppProvider';
 import { GovernanceType } from '../../../types';
 import { BarLoader } from '../../ui/loaders/BarLoader';
 import Divider from '../../ui/utils/Divider';
-import ERC20TokenContainer from './components/ERC20Token';
-import ERC721TokensContainer from './components/ERC721Token';
-import MetadataContainer from './components/Metadata';
-import { ModulesContainer } from './components/Modules';
-import SignersContainer from './components/Signers';
+import { ERC20TokenContainer } from './components/ERC20TokenContainer';
+import { ERC721TokensContainer } from './components/ERC721TokensContainer';
+import { MetadataContainer } from './components/MetadataContainer';
+import { ModulesContainer } from './components/ModulesContainer';
+import { SignersContainer } from './components/Signers/SignersContainer';
 
-export function Settings() {
+export function DAOSettings() {
   const {
     node: { safe },
     governance: { type },
