+++ src/components/pages/DaoSettings/components/ERC721TokensContainer.tsx
@@ -1,12 +1,12 @@
 import { Flex, Text, Grid, GridItem } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
-import { SettingsSection } from '..';
-import { useFractal } from '../../../../../providers/App/AppProvider';
-import { AzoriusGovernance } from '../../../../../types';
-import { DisplayAddress } from '../../../../ui/links/DisplayAddress';
-import { BarLoader } from '../../../../ui/loaders/BarLoader';
+import { useFractal } from '../../../../providers/App/AppProvider';
+import { AzoriusGovernance } from '../../../../types';
+import { DisplayAddress } from '../../../ui/links/DisplayAddress';
+import { BarLoader } from '../../../ui/loaders/BarLoader';
+import { SettingsSection } from './SettingsSection';
 
-export default function ERC721TokensContainer() {
+export function ERC721TokensContainer() {
   const { t } = useTranslation(['settings']);
   const { governance } = useFractal();
